Fix runtime permission of entrypoint.sh
In commit 1792f52 the entrypoint.sh is changed to run by autoware user.
This is done because the `gosu` command used to drop from root user to
autoware user does not work on Arm CI. However, this caused the side
effect that `usermod` is prevented from changing the UID of the autoware
user.

In this commit we revert back to running entrypoint.sh as root but drop
to autoware user using the `chroot` command.

Change-Id: Ie4a8e2f72a476e294fb98904faf2f549ffd09847
Issue-ID: SCM-1053
Signed-off-by: Liyou Zhou <liyou.zhou@arm.com>